Use argument length instead of postfix application
Instead of a postfix application instruction:

f (g x)
f g x @ @

Use an infix application with the length of the argument:

f (g x)
f 3 g 1 x

This results in a 40% speedup. I haven't investigated the exact reason
it's faster. I'm assuming it's that popping arguments from the stack
into registers is much faster now:

Instead of iterating through each instruction on the stack and tracking
applications, you can just use the argument lengths to copy regions of
the stack one after the other.

So the runtime is still linear, but linear in the number of
arguments (which is very small) instead of being linear in the size of
the arguments (which is very large if it's a data structure).

The small tradeoff is that pushing back onto the stack is a bit slower,
because some extra work is needed to calculate the argument lengths.

The algorithm for printing the stack is simpler.